Identify files in dir_in that are missing in dir_out

Returns paths from dir_in that do not exist under the same relative path in dir_out.

Usage

ngr_fs_id_missing(dir_in, dir_out, type = "file", recurse = TRUE, ...)

Arguments

dir_in

character Path to the input directory.

dir_out

character Path to the output directory.

type

character fs::dir_ls type param input. Defaults to "file"

recurse

logical fs::dir_ls recurse param input. Defaults to TRUE

...

Not used. For passing params to fs::dir_ls

Value

A character vector of input file paths that are missing from dir_out.

See also