Filter Rows Based on String Detection in a Column Detected with a partial String Match

This function filters rows in a data frame where a specified column's name matches a pattern and the values of that column contain a specified string.

Usage

ngr_str_df_detect_filter(x, col_filter, str_filter)

Arguments

x

data.frame The input data frame to filter.

col_filter

character A single string representing the pattern to match column names.

str_filter

character A single string to match within the values of the filtered column.

Value

A filtered data frame if matching rows are found; otherwise, the original data frame is returned.

See also