Create an HTML Link to Repository Resources or GitHub Pages

This function generates an HTML <a> tag linking to a specified repository resource or GitHub Pages site hosted from the repository. It is particularly useful for embedding links in HTML documents.

Usage

ngr_str_repo_link(
  repo_source,
  url = "https://www.newgraphenvironment.com",
  anchor_text = "url_link",
  target = "_blank"
)

Arguments

repo_source

character A single string representing the repository name to be linked.

url

character A single string specifying the base URL for the repository host. Default is "https://www.newgraphenvironment.com".

anchor_text

character A single string specifying the text displayed for the link. Defaults to "url_link".

target

character A single string indicating the target attribute in the HTML link. Default is "_blank". Options include:

"_blank": Opens the link in a new tab or window.

"_self": Opens the link in the same tab or window.

"_parent": Opens the link in the parent frame, if applicable.

"_top": Opens the link in the full body of the window, ignoring frames.

Value

character A single string containing the HTML <a> tag.

Details

This function can create links to either:

Repository resources: Links directly to the repository on platforms like GitHub.

GitHub Pages: Links to websites or documentation served from the repository, using a base URL for the hosted site.

If the provided url does not start with http:// or https://, a warning is issued via cli::cli_alert_warning() to indicate potential issues with the generated links.

See also

Examples

# Example 1: Link to the repository with default anchor text
ngr_str_repo_link("ngr", url = "https://github.com/NewGraphEnvironment")
#> [1] "<a href=\"https://github.com/NewGraphEnvironment/ngr\" target=\"_blank\">url_link</a>"

# Example 2: Link to GitHub Pages with anchor text set to the repository name
ngr_str_repo_link("ngr", url = "https://www.newgraphenvironment.com", anchor_text = "ngr")
#> [1] "<a href=\"https://www.newgraphenvironment.com/ngr\" target=\"_blank\">ngr</a>"